Simplify config structure
